requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
pydantic>=1.10,<2.0
python-dotenv>=1.0.0,<2.0.0
llama-parse>=0.4.0
openai>=1.0.0

